Proper place the helper methods into a module
